Added Modal to the Delete User Button

+++ client/src/components/Dashboard/UserTable/UserTable.jsx
@@ -5,32 +5,51 @@ import { AccountBox, Delete } from "@material-ui/icons";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
 import { deleteUser, getUsers } from "../../../redux/apiCalls";
+import { Box, Modal, Typography } from "@mui/material";
+import { useState } from "react";
 
 const UserTable = () => {
   const dispatch = useDispatch();
   const users = useSelector((state) => state.users.users);
-
+  const [id, setId] = useState();
+  const [firstName, setFirstName] = useState();
+  const [lastName, setLastName] = useState();
+  const [isAdmin, setIsAdmin] = useState(false);
   useEffect(() => {
     getUsers(dispatch);
   }, [dispatch]);
 
   const handleDelete = (id) => {
-    deleteUser(id, dispatch);
+    if (!isAdmin) {
+      deleteUser(id, dispatch);
+    }
   };
 
   const columns = [
-    { field: "_id", headerName: "ID", width: 220 },
     {
-      field: "user",
-      headerName: "User",
+      field: "_id",
+      headerName: "ID",
+      width: 300,
+    },
+    {
+      field: "firstName",
+      headerName: "First Name",
       width: 200,
       renderCell: (params) => {
         return (
           <div className={styles.userListItem}>
             <img className={styles.userListImg} src={params.row.image} alt="" />
-            {params.row.firstName + " " + params.row.lastName}
+            {params.row.firstName}
           </div>
         );
+      },
+    },
+    {
+      field: "lastName",
+      headerName: "Last Name",
+      width: 200,
+      renderCell: (params) => {
+        return <div className={styles.userListItem}>{params.row.lastName}</div>;
       },
     },
     {
@@ -40,21 +59,29 @@ const UserTable = () => {
       renderCell: (params) => {
         return (
           <div className={styles.userListItem}>
-            <img className={styles.userListImg} src={params.row.image} alt="" />
             {params.row.createdAt.split("T", 1)}
           </div>
         );
       },
     },
     {
+      field: "signedIn",
+      headerName: "Signed In At",
+      width: 200,
+      renderCell: (params) => {
+        return <div className={styles.userListItem}>{params.row.signedIn}</div>;
+      },
+    },
+    {
       field: "email",
       headerName: "Email",
-      width: 160,
+      width: 200,
     },
+
     {
       field: "action",
       headerName: "Action",
-      width: 150,
+      width: 200,
       renderCell: (params) => {
         return (
           <div className={styles.actions}>
@@ -69,7 +96,13 @@ const UserTable = () => {
             </Link>
             <button
               className={styles.deleteButton}
-              onClick={() => handleDelete(params.row._id)}
+              onClick={() => {
+                setId(params.row._id);
+                setFirstName(params.row.firstName);
+                setLastName(params.row.lastName);
+                setIsAdmin(params.row.isAdmin);
+                setOpen(true);
+              }}
             >
               <Delete className={styles.deleteIcon} />
             </button>
@@ -78,8 +111,49 @@ const UserTable = () => {
       },
     },
   ];
+  const [open, setOpen] = useState(false);
+
+  const handleClose = () => setOpen(false);
   return (
     <div className={styles.datatable}>
+      <Modal
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="modal-modal-title"
+        aria-describedby="modal-modal-description"
+      >
+        <Box className={styles.box}>
+          <div className={styles.modalHeader}>
+            <Typography
+              id="modal-modal-title"
+              variant="h6"
+              component="h2"
+              style={{ fontWeight: "600", color: "#001219" }}
+            >
+              Are you sure?
+            </Typography>
+          </div>
+          <div className={styles.modalBody}>
+            <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+              The user ({firstName} {lastName}) will be deleted.
+            </Typography>
+          </div>
+          <div className={styles.modalFooter}>
+            <button className={styles.discard} onClick={handleClose}>
+              Discard
+            </button>
+            <button
+              className={styles.delete}
+              onClick={() => {
+                handleDelete(id);
+                handleClose();
+              }}
+            >
+              DELETE
+            </button>
+          </div>
+        </Box>
+      </Modal>
       <div className={styles.datatableTitle}>User List</div>
       <DataGrid
         rows={users}
@@ -88,7 +162,6 @@ const UserTable = () => {
         getRowId={(row) => row._id}
         pageSize={9}
         disableSelectionOnClick
-        checkboxSelection
       />
     </div>
   );
